move core files to the core folder
