+++ src/core/App.js
@@ -1,5 +1,5 @@
-import Form from "./features/Form";
-import Clock from "./features/Clock/clock.js";
+import Form from "../features/Form";
+import Clock from "../features/Clock/clock.js";
 
 function App({ result, calculateResult }) {
   return (
