+++ src/core/GlobalStyle.js
@@ -1,5 +1,5 @@
 import { createGlobalStyle } from "styled-components";
-import background from "./assets/background.png";
+import background from "../assets/background.png";
 
 export const GlobalStyle = createGlobalStyle`
   html {
